chore: update storybook version
