+++ app/client/packages/design-system/widgets/src/components/Button/Button.stories.mdx
@@ -1,4 +1,4 @@
-import { Canvas, Meta, Story, ArgsTable } from "@storybook/addon-docs";
+import { Canvas, Meta, Story, ArgsTable } from "@storybook/blocks";
 
 import { Button } from "./";
 import { Icon } from "@design-system/headless";
