+++ app/client/packages/storybook/.storybook/manager-head.html
@@ -1,74 +1,8 @@
 <style>
-
-  div[role="main"] {
-    border: 1px solid #E3E8EF !important;
-  }
-
-  div[role="main"] > div {
-    box-shadow: none !important;
-  }
-
   img[alt="Appsmith Design System"] {
     width: 134px;
   }
-
-  .sidebar-item[data-selected="true"] {
-    background-color: #E3E8EF !important;
-    color: #364252 !important;
-    font-weight: 500 !important;
-  }
-
-  .sidebar-item[data-selected="true"] > svg{
-    color: #364252 !important;
-  }
-
-  .sto-rfiy1p, .sto-f4xokn {
-    border-top: 2px solid transparent !important;
-    border-bottom: 2px solid transparent !important;
-    font-weight: 500 !important;
-  }
-
-  .tabbutton:focus {
-    border-bottom-color: #F86A2B !important;
-  }
-
-  .sto-f4xokn {
-    border-bottom-color: #F86A2B !important;
-  }
-
   #storybook-explorer-tree > div {
     margin: unset !important;
-  }
-
-  .sidebar-item {
-    padding-top: 6px !important;
-    padding-bottom: 6px !important;
-  }
-
-  #storybook-explorer-searchfield, button[title="Shortcuts"] {
-    border-radius: 0 !important;
-    border: 1px solid #E3E8EF !important;
-    box-shadow: none !important;
-  }
-
-  button[title="Shortcuts"]:focus:before {
-    border-radius: 0 !important;
-    border: 1px solid #E3E8EF !important;
-  }
-
-  .sto-1jd7xzf:focus,
-  .sto-qbvvbz input:focus,
-  .sto-dtdhmd:focus {
-    box-shadow: #F86A2B 0 0 0 1px inset !important;
-  }
-
-  .sto-t2643i:hover,
-  .sto-t2643i:focus-visible,
-  .sto-1k5e3f:hover,
-  .sto-1k5e3f:focus-visible,
-  .sto-14xfrgt:hover,
-  .sto-14xfrgt:focus-visible {
-    color: #F86A2B !important;
-    background: #f86a2b1f !important;
   }
 </style>
