+++ app/client/packages/storybook/.storybook/preview.js
@@ -3,18 +3,37 @@ import "@blueprintjs/icons/lib/css/blueprint-icons.css";
 import "@blueprintjs/core/lib/css/blueprint.css";
 import "./styles.css";
 
-export const parameters = {
-  actions: { argTypesRegex: "^on[A-Z].*" },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
+export const decorators = [theming];
+
+const customViewports = {
+  kindleFire2: {
+    name: "Kindle Fire 2",
+    styles: {
+      width: "600px",
+      height: "963px",
     },
   },
-  backgrounds: {
-    disable: true,
+  kindleFireHD: {
+    name: "Kindle Fire HD",
+    styles: {
+      width: "533px",
+      height: "801px",
+    },
   },
-  layout: "centered",
 };
 
-export const decorators = [theming];
+const preview = {
+  globalTypes: {
+    colorMode: {},
+    borderRadius: {},
+    accentColor: {},
+    fontFamily: {},
+    rootUnitRatio: {},
+  },
+  parameters: {
+    viewport: { viewports: customViewports },
+    actions: { argTypesRegex: "^on[A-Z].*" },
+  },
+};
+
+export default preview;
