+++ app/client/packages/storybook/.storybook/styles.css
@@ -1,19 +1,8 @@
 html,
 body,
-#root {
+#storybook-root {
   height: 100%;
   width: 100%;
-}
-
-*,
-:after,
-:before {
-  border: 0 solid #e4e4e7;
-  box-sizing: border-box;
-}
-
-.innerZoomElementWrapper > * {
-  overflow: hidden;
 }
 
 /* fonts */
