+++ app/client/packages/storybook/package.json
@@ -5,8 +5,8 @@
   "author": "Valera Melnikov <valera@appsmith.com>, Pawan Kumar <pawan@appsmith.com>",
   "license": "MIT",
   "scripts": {
-    "storybook": "start-storybook -p 6006",
-    "build": "build-storybook",
+    "storybook": "storybook dev -p 6006",
+    "build": "storybook build",
     "chromatic": "chromatic --project-token CHROMATIC_PROJECT_TOKEN"
   },
   "devDependencies": {
@@ -14,27 +14,27 @@
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-react": "^7.18.6",
     "@babel/preset-typescript": "^7.21.0",
-    "@storybook/addon-actions": "^6.5.16",
-    "@storybook/addon-essentials": "^6.5.16",
-    "@storybook/addon-interactions": "^6.5.16",
-    "@storybook/addon-postcss": "^2.0.0",
-    "@storybook/builder-webpack5": "^6.5.16",
-    "@storybook/manager-webpack5": "^6.5.16",
-    "@storybook/node-logger": "^6.5.16",
-    "@storybook/preset-create-react-app": "^4.1.2",
-    "@storybook/react": "^6.5.16",
-    "@storybook/testing-library": "^0.0.13",
-    "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
-    "autoprefixer": "^9.0.0",
+    "@storybook/addon-actions": "^7.1.1",
+    "@storybook/addon-controls": "^7.1.1",
+    "@storybook/addon-docs": "^7.1.1",
+    "@storybook/addon-measure": "^7.1.1",
+    "@storybook/addon-outline": "^7.1.1",
+    "@storybook/addon-toolbars": "^7.1.1",
+    "@storybook/addon-viewport": "^7.1.1",
+    "@storybook/blocks": "^7.1.1",
+    "@storybook/components": "^7.1.1",
+    "@storybook/manager-api": "^7.1.1",
+    "@storybook/preset-create-react-app": "^7.1.1",
+    "@storybook/react": "^7.1.1",
+    "@storybook/react-webpack5": "^7.1.1",
+    "@storybook/theming": "^7.1.1",
     "babel-loader": "9.1.2",
     "chromatic": "^6.19.8",
-    "postcss": "^8.4.23",
-    "storybook-color-picker": "^3.1.0",
+    "react-docgen-typescript": "^2.2.2",
+    "storybook": "^7.1.1",
     "tsconfig-paths-webpack-plugin": "^4.0.1"
   },
   "dependencies": {
-    "colorjs.io": "^0.4.3",
-    "react-docgen-typescript": "^2.2.2"
+    "colorjs.io": "^0.4.3"
   }
 }
